requirements.txt
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
contourpy==1.3.3
cycler==0.12.1
fast-plate-ocr==1.1.0
filelock==3.25.2
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2026.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
kiwisolver==1.5.0
lap==0.5.13
lxml==6.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
onnxruntime==1.24.4
onvif_zeep==0.2.12
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==12.2.0
platformdirs==4.9.6
polars==1.39.3
polars-runtime-32==1.39.3
protobuf==7.34.1
psutil==7.2.2
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.7
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==15.0.0
scipy==1.17.1
setuptools==81.0.0
six==1.17.0
sympy==1.14.0
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
typing_extensions==4.15.0
ultralytics==8.4.37
ultralytics-thop==2.0.18
urllib3==2.6.3
zeep==4.3.2
